K8S-502: Avoid listing volumes on Pod deletion unless specified by config
When a Pod is deleted the operator will always try to list any volumes associated
 with the Pod.  This is unnecessary, and in the case where security doesn't allow
 the listing of volumes this can result in errors masking underlying reason
 for failure of Pod creation.

Reviewed-on: http://review.couchbase.org/97702
Tested-by: Build Bot <build@couchbase.com>

+++ pkg/util/k8sutil/pod_util.go
@@ -245,7 +245,8 @@ func podVolumeSpecForClaim(configName, claimName string) v1.Volume {
 }
 
 // Delete pod and any associated persisted volumes
-func DeleteCouchbasePod(kubeCli kubernetes.Interface, namespace, clusterName, name string, opts *metav1.DeleteOptions) error {
+// when removeVolumes is 'true'
+func DeleteCouchbasePod(kubeCli kubernetes.Interface, namespace, clusterName, name string, opts *metav1.DeleteOptions, removeVolumes bool) error {
 
 	var errs []string
 
@@ -253,8 +254,10 @@ func DeleteCouchbasePod(kubeCli kubernetes.Interface, namespace, clusterName, na
 		errs = append(errs, err.Error())
 	}
 
-	if err := deletePodVolumes(kubeCli, namespace, clusterName, name); err != nil {
-		errs = append(errs, err.Error())
+	if removeVolumes {
+		if err := deletePodVolumes(kubeCli, namespace, clusterName, name); err != nil {
+			errs = append(errs, err.Error())
+		}
 	}
 
 	if len(errs) > 0 {
